requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-gis==1.2.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
djangorestframework-simplejwt